@@ -10,30 +10,38 @@
     "luxon": "^3.4.3",
     "match-sorter": "^6.3.1",
     "prop-types": "^15.8.1",
-    "react": "^17.0.1",
+    "react": "^18.2.0",
     "react-dnd": "^14.0.5",
     "react-dnd-html5-backend": "^14.1.0",
-    "react-dom": "^17.0.1",
+    "react-dom": "^18.2.0",
     "react-intersection-observer": "^8.32.5",
     "react-redux": "^7.2.9",
-    "react-scripts": "5.0.1",
+    "react-scripts": "^5.0.1",
     "redux": "^4.2.1",
     "redux-devtools-extension": "^2.13.9",
     "redux-thunk": "^2.4.2",
-    "sort-by": "^1.2.0",
+    "sort-by": "^0.0.2",
+    "uuid": "^9.0.1",
     "web-vitals": "^1.0.1"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
+    "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.1.10",
     "@types/luxon": "^3.3.3",
     "@types/react": "17.0.0",
-    "@types/react-dom": "17.0.0",
+    "@types/react-dom": "^18.2.15",
+    "@types/uuid": "^9.0.7",
+    "cypress": "^13.4.0",
+    "fetch-mock": "^9.11.0",
+    "jest": "^27.5.1",
     "react-router-dom": "^6.17.0",
+    "react-test-renderer": "^18.2.0",
+    "redux-mock-store": "^1.5.4",
     "typescript": "^4.9.5"
   },
   "scripts": {
+    "cypress:open": "./node_modules/.bin/cypress open",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
