@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from 'react-dom/client';
 import "./index.css";
 import App from "./app";
 import reportWebVitals from "./reportWebVitals";
@@ -7,15 +7,17 @@ import { Provider } from 'react-redux';
 import { store } from "./services";
 import {BrowserRouter} from "react-router-dom";
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container!);
+
+root.render(
   <React.StrictMode>
       <Provider store={store}>
           <BrowserRouter>
             <App />
           </BrowserRouter>
       </Provider>
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
